Remove compilation warnings

+++ ion/include/ion/unicode/utf8_decoder.h
@@ -28,7 +28,7 @@ public:
   CodePoint previousCodePoint();
   const char * stringPosition() const { return m_stringPosition; }
   static size_t CharSizeOfCodePoint(CodePoint c);
-  static size_t CodePointToChars(CodePoint c, char * buffer, int bufferSize);
+  static size_t CodePointToChars(CodePoint c, char * buffer, size_t bufferSize);
 private:
   const char * const m_string;
   const char * m_stringPosition;
